Tasks are now users bound

+++ backend/src/controllers/tasksController.ts
@@ -2,10 +2,17 @@ import { RequestHandler } from "express";
 import TaskModel from "../models/task"
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefined } from "../util/assertIsDefined";
 
 export const getTasks: RequestHandler = async (req, res, next) => {
+
+    const authenticatedUserId = req.session.userId;
+
     try {
-        const tasks = await TaskModel.find().exec();
+        assertIsDefined(authenticatedUserId);
+
+
+        const tasks = await TaskModel.find({userId: authenticatedUserId}).exec();
         res.status(200).json(tasks)
 
     } catch (error) {
@@ -22,8 +29,10 @@ export const createTask: RequestHandler<unknown, unknown, CreateNoteBody, unknow
     const taskName = req.body.taskName;
     const isImportant = req.body.isImportant
 
-    try {
+    const authenticatedUserId = req.session.userId;
 
+    try {
+        assertIsDefined(authenticatedUserId);
         //check if taskName is defined
         if (!taskName) {
             throw createHttpError(400, "Task must have a name")
@@ -31,7 +40,9 @@ export const createTask: RequestHandler<unknown, unknown, CreateNoteBody, unknow
 
         const newTask = await TaskModel.create({
             name: taskName,
-            isImportant: isImportant
+            isImportant: isImportant,
+            userId: authenticatedUserId
+            
         });
         res.status(201).json(newTask);
     } catch (error) {
@@ -40,9 +51,13 @@ export const createTask: RequestHandler<unknown, unknown, CreateNoteBody, unknow
 };
 
 export const getTask: RequestHandler = async (req, res, next) => {
+
+    const authenticatedUserId = req.session.userId;
+
     const taskId = req.params.taskId;
 
     try {
+        assertIsDefined(authenticatedUserId);
         // check if valid id signature
         if (!mongoose.isValidObjectId(taskId)) {
             throw createHttpError(400, "Invalid task id");
@@ -53,6 +68,10 @@ export const getTask: RequestHandler = async (req, res, next) => {
         // check if task null or undefined
         if (!task) {
             throw createHttpError(404, "Task not found");
+        }
+
+        if(!task.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "You dont have access to this info")
         }
 
         res.status(200).json(task);
@@ -78,9 +97,11 @@ export const updateTask: RequestHandler<UpdateTaskParams, unknown, UpdateTaskBod
     const newTask = req.body.taskName
     const isCompleted = req.body.isCompleted
     const isImportant = req.body.isImportant
+    const authenticatedUserId = req.session.userId;
 
 
     try {
+        assertIsDefined(authenticatedUserId);
         // check if valid id signature
         if (!mongoose.isValidObjectId(taskId)) {
             throw createHttpError(400, "Invalid task id");
@@ -95,6 +116,10 @@ export const updateTask: RequestHandler<UpdateTaskParams, unknown, UpdateTaskBod
         // check if task null or undefined
         if (!task) {
             throw createHttpError(404, "Task not found");
+        }
+
+        if(!task.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "You dont have access to this info")
         }
 
         task.name = newTask;
@@ -112,18 +137,24 @@ export const updateTask: RequestHandler<UpdateTaskParams, unknown, UpdateTaskBod
 
 export const deleteTask: RequestHandler = async (req, res, next) => {
     const taskId = req.params.taskId
+    const authenticatedUserId = req.session.userId;
 
     try {
+        assertIsDefined(authenticatedUserId);
         // check if valid id signature
         if (!mongoose.isValidObjectId(taskId)) {
             throw createHttpError(400, "Invalid task id");
         }
-
+        
         const task = await TaskModel.findById(taskId).exec()
 
         // check if task null or undefined
         if (!task) {
             throw createHttpError(404, "Task not found");
+        }
+
+        if(!task.userId.equals(authenticatedUserId)){
+            throw createHttpError(401, "You dont have access to this info")
         }
 
         await task.deleteOne();
